porting layer: Fix wrong calculation of iterative mean
The code that calculates the iterative mean of execution times uses the
loop variable i as the number of measurements. This is incorrect, as the
first iterations is ignored on some tests. Furthermore, if the number of
cycles measured is smaller than a gap value, the loop iteration's
measurement is ignored. Both cases desync the number of measured values
from the loop variable i. Add a variable num_measurements which always
holds the correct number of measurements added to the iterative mean in
order to fix this.